apparmor: allow open of /etc/mosquitto/conf.d
The apparmor profile allowed read of files under /etc/mosquitto/conf.d, but not the directory itself. When include_dir is given to load the files under the directory, an attempt to open the directory with openat() is made, which fails under apparmor unless the directory is explicitly marked as readable.